auth_server/requirements.txt
SQLAlchemy==1.4.42
pytest==7.2.0
email-validator==1.3.0
PyJWT==2.6.0
